+++ docs/setup/release-notes.md
@@ -66,6 +66,7 @@
 * [<a href='https://github.com/apache/sedona/issues/3281'>GH-3281</a>] - Implement `GeoDataFrame.from_features` for in-memory GeoJSON-like features
 * [<a href='https://github.com/apache/sedona/issues/3288'>GH-3288</a>] - Implement `GeoDataFrame.from_dict` for in-memory dictionaries
 * [<a href='https://github.com/apache/sedona/issues/3344'>GH-3344</a>] - Expose `SpatialIndex.valid_query_predicates`
+* [<a href='https://github.com/apache/sedona/issues/3357'>GH-3357</a>] - Implement `list_layers` for GeoPackage using the native Spark reader
 
 ### Bug Fixes